refactor: move map component
